Initial attempt at GitHub Actions for cfgov-refresh
This change ports out Travis CI jobs to GitHub Actions as faithfully as possible. It removes the custom scripts and moves all the necessary commands and context to .github/workflows. There's probably refactoring that could be done, and we may want to run our jobs differently on GitHub than we do on Travis, but this is intended only to migrate and get us started.

One change I've made from Travis is to break out the Python linting from the unit tests.